Changed AddTask.js
This js now does everything starting from getting the variable from AutoNotication in
 then splitting it to get various parts of the message (AN takes care to send
only eligible messages forward).

Then we make some ajax calls to add the task and its comment and depending on the
result of the ajax call, generate the sms_text variable. This is then used by
AutoNotification's reply action to automatically send the status of the task back.